requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
ez-yaml==2.2.0
    # via -r requirements.in
pyrage==1.2.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via yamlpath
ruamel-yaml==0.17.21
    # via
    #   ruamel-yaml-string
    #   yamlpath
ruamel-yaml-string==0.1.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
yamlpath==3.8.2
    # via -r requirements.in
